[CCode] Fix for writing divided arrays into CCode files
Arrays containing more than 400 elements are divided into smaller parts when writing to CCode. In the case of i.e. vkDependencyInfo struct, it contains data about three arrays. The error was that, for example, two arrays needed to be split, and after the first split the number of elements in one array was less than 400 elements and the other was bigger that 400 and still needed to be split. During the partial write, the check in the function continued to be performed on the basis of the overall size of the array and not the current position of the write. In each successive iteration of dividing the larger of the two arrays, it caused an attempt to also write an array whose elements had already been written in earlier iterations, resulting in a writing zero-size array, instead of nullptr. In the given solution, we make the partial write independent of the overall length of the array and rely only on the given interval to be written.