[OGL] Use correct size in glClearBufferSubData
Fixes a harmless bug in glClearBufferSubData, where size of the buffer
slice to be cleared was passed to CBinaryResource constructor instead of
size of the texel pointed to by `data`. The texel size in bytes differs
based on the `type` and `format` parameters. Surprisingly, reading too
much data did not cause any crashes, it resulted only in storing too
much data in the stream.