[common] Fix for restoration timer
Fixed bug: count and display of state restore time in streams without state restore. Description: The restoration timer was initialized with the paused variable set to false, which should have been set to true. The constructor as well as the Timer::Start() function have also been changed. If the constructor is called with paused set to true, the start function should not be called. In the Start() function, the paused variable is set to false; without this, a timer that was set to paused from the beginning of its operation and then the start function was called on it would start counting the time, while it would continue to be treated as paused in the other functions.