[common] Fix playback timer
Fixed bug: Displaying the playback time equal to the total program time, not the actual playback. Description: The start for the playback timer was called in the switch for case: INIT_END, which is only in streams with state restore. This caused the time to be counted from the initialization of the Timer, not from the actual playback. A one-time start in case for FRAME_START was added, as well as the timer was disabled by default by initializing it as paused.