[OGL][CCode] Fix potential char* use after free
For some reason we were converting the std::filesystem::path to
std::string and taking a char* from it (and later converting it back to a
path). When that char* was later used, it might have already been
deallocated, as the string is a temporary.

Now we just pass the path directly instead of converting it to a string
only to convert it back to a path.

The same issue was present in multiple places, all four of them were
fixed.