[Vk] Start using the Token class in generator
Also remove code for handling the `recCond` and `preSchedule`
annotations on tokens, because they were not being used. Remove
inheritance support too, as it was not used in Vk.

The `custom` annotation is also not used on tokens (only on structs),
but it's kept for now to make the diff more readable (its removal
changes a lot of indentation). It will be removed in a separate commit.