[OGL] Wrap generator's bitflags in an IntFlag
In generator, we had a bunch of separate variables representing various
function types as bits, used in a bitflag-like manner. (Bind, Fill,
Param, etc.) Now they are wrapped in an IntFlag - a subclass of Enum and
int, meant specifically for bitflags with specified bit numbers.

We also convert generator data to use `FuncType` explicitly, e.g.,
`FuncType.Get` instead of just `Get`.

For compatibility with existing code, we use the @global_enum annotation
to export bit names into the global namespace. In the future, all code
should be changed to not require this compatibility feature. The
annotation could then be removed.