[OGL] Add a new generator using Mako templates
The legacy generator is still used by default. The new one generates
files into a `temp/` folder - this way output of both generators can be
compared to see if the new one is working properly. And it is - output
is the same except for small whitespace differences due to improved
formatting and updated year in license headers.