[OGL] Replace arg1, arg2, etc. with args list
Instead of explicitly numbering arguments and just putting each of them
separately as a piece of function metadata, we now group them all in a
Python list, listing them in order.

This means we don't need extra code to collect them into a list later.

This commit also ensures that each line ends with a comma (including
tokens with no arguments) to unify the style between lines, allow for
easier updating of data through a script, and give a hint to formatters
to put each argument on a separate line.