[Vk] Add a Mako template for vulkanLogAuto.inl
This commit adds code modifying generator input on the fly to a format
suitable for the new generator. (By stripping '_' from struct names.)
Old generator got code for doing the reverse, to keep the input it used
to get. If not for this compatibility code in old generator, its output
files would differ from intended ones.