[Vk] Add a Mako template for vulkanFunctions.cpp
And supporting generate_vk.py code, of course.

Notable change is in how we strip types to their essentials (e.g., from
`const float*` to just `float`). We used to remove "const " as a part of
type stripping, but this failed on types like `const
VkAccelerationStructureBuildRangeInfoKHR* const*` where there is no
space after one of the consts. This commit changes that removed string
to simply "const" (without the space).