[Vk] Add vulkanArgumentsAuto.cpp.mako
The otput differs a bit from the old generator - useless attempts at
extraction of mapped pointers from primitive type arrays (e.g.,
`Cuint32_t::CSArray`) are omitted.
`CArgumentSizedArray::GetMappedPointers()` returns an empty set anyway,
since contents of this array are primitive types that are not mapped.

For now we match the behavior of old generator in that we don't try to
extract mapped pointers from structures potentially hiding behind
`pNext` pointers. This will likely need to change in the future.

Additional logic was added to the `fields_to_str` helper function. It
allows generating proper constructor arguments for each field of a
struct.