[Vk] Fix IntFlag.__len__/__iter__ in Python < 3.11
Since Python 3.11, instances of Flag (and Flag subclasses) are iterable
and have length. We want this functionality in older Python versions, so
we write it ourselves and use an admittedly ugly hack to add __len__ and
__iter__ to IntFlag if they are missing. We also replace existing
__len__/__iter__ if they are broken. (Python 3.10 on Windows has
IntFlag.__iter__ that does the wrong thing.)

Please revert this commit when Python 3.10 and older are no longer
supported.