[Vk] Move VkFlags typedefs to a generated file
We used to have the VkFlags/VkFlags64 typedefs in vulkan_basic.h, but
there's no need to keep them checked into Git. We can recreate them from
generator data. We do so in vulkanHeader.h.

We also change VkDescriptorBindingFlagsEXT to VkDescriptorBindingFlags
in one file, as our generator only handles the latter, but doesn't
create a typedef for the EXT name.