[VK] Fix problems reported by validation layers with new functions
Validation layers complained that when a flag is specified for queue
during a device creation, it needs to be acquired with flag-capable
function vkGetDeviceQueue2(). So far it was done only for queues that
actually had any flags specified. But validation layers complain in such
situation that if a queue is specified twice during device creation
(once without any flags and once with flags), vkGetDeviceQueue2() needs
to be used in both situations (for both queue handles).

So I think we should just use the function whenever it is available
and can be called.