[OGL] Fix a crash during subcapture
Add missing uniform types to CVariableGLSLInfo::UniformDimensions. Group
and sort them to make code clearer. Log an error and throw when
encountering an unknown uniform type. (Previously, that unknown type was
inserted into the map with a null value, which was then returned. The
null value caused a crash in turn.)