[DX12] Split Trace layer into smaller layers
This commit removes the (unnecessary) dependencies between:

- Trace layer
- Debug layer(s)
- Conversion of enums and GUIDs to string
- Fast IO

This was originally grouped as such because the Debug layer(s) write to
the trace files (using FastOStream) and this created a dependency. Also,
the Debug layer benefited from the "print" functionality of Trace.

Using messages we can detangle this dependencies (the Debug layer sends
a message with the Trace layer as a subscriber).

This allows us to remove the FastOStream dependencies from the Debug
layer(s) and also provide the helper functions to print enums and GUIDs
which are useful to plugins and other parts of GITS.

The commit also fixes small bugs with the Debug and Trace layers.