[VK] Make SBT patching enabled by default
This commit makes shader group handles patching to be enabled by default.
Original option to enable it is replaced with an option to force disable
SBT patching.

Each pipeline is checked if its handles are different than
those used during recording. If they are identical, patching is not used
for that particular pipeline.

Additionally, this commit improves checking if a GITS recorder is attached
a player. Previous implementation treated other Vulkan loader libraries as
a recorder too. Introduced a custom function which is exported only by the
recorder, to distinguish between it and other Vulkan loader libraries.