[DX12] Subcapture/ResourceDump/RtasCache - handle all fence signals
If all types of signals on fence are not communicated to the GpuExecutionTracker then it will think that some of the queues are stuck on Wait and will not mark executables as ready and waitUntilDumped will not wait on them.
The most relevant change is handling EnqueueMakeResident as it can change fence value.