[DX12] Simplified the SwapChain buffers tracking
Initially GITS was acquiring all the buffers (using GetBuffer) on
SwapChain creation. I am not sure why this was needed for (?) but it had
many side effects. In particular it increased the reference count of the
SwapChain and also required special tracking and management of
ResizeBuffers.

This commit simplifies this by removing the unnecessary (?) acquire and
releaseBuffers calls during capture time. This also helps us simplify
the tracking required by subcapture.