[DX12] Fixed SwapChain buffer tracking in recorder
It is possible for an application to call GetBuffer multiple times with
the same index to retrieve the same back buffer.

This caused the std::vector to contain repeated interfaces (which
creates an issue on release). The fix is to use an std::unordered_set
instead of an std::vector.