[DX12] Refactor the IDXGIAdapter override logic
Previously the adapter override check was part of the D3D12CreateDevice
in replayCustomizationLayer.cpp. This was not sufficient because the
IDXGIAdapter may be required before D3D12 device creation. For example
as part of the process to load Intel Extensions.

This change moves the adapter override functionality to AdapterService
and also adds extra print outs into the log (or console) for clarity.

Changes:
- Moved the adapter override logic into AdapterService (in PlayerManager)
- Added call to get a "default adapter". This is used by Intel Extensions
and also on device creation when pAdapter is set to NULL
- PlayerManager loads the adapter override right after loading runtime
libraries
- Added toStr(wchar_t*) to avoid duplication