[VK] Disable AS memory aliasing check
Some games create and build acceleration structures in long dependency chains
where AS can be built, copied to another AS, which is then a source of an
update operation to yet another AS, which is then compacted to yet another AS.
Some acceleration structures in this chain (probably) share parts of memory
which require acceleration structures to be restored in the exactly same order
but, at the same time, some of the operations cannot be skipped (it's not
enough to restore only the last acceleration structure).

This commit brings back building all the acceleration structures in a state
restore phase, even those who share/alias memory.