[common] Refactor CTokenFrameNumber
CTokenFrameNumber is used for multiple markers (not just frame start /
end). This change refactors the usage of CTokenFrameNumber into a more
intuitive CTokenMarker which is used to mark common sections of the
GITS stream.

This change also fixes an issue where the Size() returned was incorrect
and (also set to 0 for all tokens).