[CCode] Get rid of OpenGL includes in helper.h
The openglDrivers.h include was moved to helperGL.h. This meant helper.h
no longer knew about GLubyte, which it still used. An example OpenGL
CCode built just fine with the `operator const GLubyte*()` removed, but
CCode from other apps could still be affected, it's best to play it
safe. Since GLubyte is usually a typedef to unsigned char, we replace it
with that.