[OGL] Fix "Couldn't map uniform block index" warn
This warning is caused by glUniformBlockBinding calls made by player.
When recording substreams, these calls are saved in the substream as if
they were made by the app. But they were not present in the original
stream.

The mapping fails because we add map entries on
glGetProgramResourceIndex and glGetUniformBlockIndex calls. But unlike
apps, the player does not call any of these before calling
glUniformBlockBinding, so the map is empty when we try to read the index
from it.

This commit makes the player call glGetUniformBlockIndex before
glUniformBlockBinding. The result of that get is not important and it's
not used anywhere. All that matters is that the call gets recorded in
the substream.

The proper fix would be to not save calls originating from the player to
substreams. But currently we lack the infrastructure to distinguish
calls from the stream from calls injected by the player.