[common][DX12] Added new token with a UInt64 value
The new token CTokenMarkerUInt64 can be used to mark the stream with a
uint64 value. This is a common CToken that may get used by both the common
framework as well as any given API.

This commit also implements the D3D12 support for this token. A new
command MarkerUInt64Command is added to the layer interface. The first
layer to use markers is "PrintStatus" which simply prints the time for
each state restoration phase.