[config] Generate negative flags for bool options
Previously there was no way to set bool config values to false from the
command line. This commit gives users the option to do it by introducing
"negative" argument for each bool argument (prefixed with "no").
This approach was chosen to not change the command line interface at this time.