[VK] Enable tracing of all structs by default
Simple tracing is rarely used in Vulkan. Most of the time showing values
of members of all structures is crucial for debugging purposes. That's
why it's inconvenient to always add a command-line option to enable
tracing of structure members.

This commit removes the option entirely and enables full tracing all
the time. As an added benefit, it simplifies the generated code and
improves compilation times.