[common] Remove CCode support
This is the first of a series of commits that will aim to remove CCode
support from GITS.

This initial commit:
- Removes CCode mentions from the documentation
- Removes CCode files
- Removes CCode install steps from CMakeList files
- Removes CCode from the configuration

The full removal of CCode from the codebase will be done in multiple
commits. This is because CCode is very integrated into the GITS core and
also in the different backends.

Removing everything related to CCode in a single PR would be a lot of
work at once.