[common] Make plog respect LogToConsole setting
This commit creates separate LogToConsole settings for player and
recorder, while maintaining the shared LogToConsole setting, which is
now derived from the corresponding recorder/player setting. The shared
setting means checking for LogToConsole is uniform in all sorts of
shared code.

LogToConsole value in the config file defaults to true in player, false
in recorder. It's assumed to be true before config parsing - GITS logs
to console from the start and only stops when it realizes that
LogToConsole is set to false in config. It's done this way because if it
was off by default and config parsing failed (or another early error
happened), the error message wouldn't be printed to the console. Yes, it
means we might print something to the console contrary to user's will,
but it shouldn't be a problem, as the main use case for disabling
console logging is to avoid the performance hit that comes with high
volume logging (for example, when log level is set to TRACE and we are
already logging to a file, so no need to log to console). In the future,
we can consider accumulating logs in memory until we parse the config
and know where user wishes us to print the logs. The volume of this
early logging should be very small, so keeping it in memory shouldn't be
a problem.