[third_party] Allow control of the Vulkan Dear ImGui patch
Previous version of the patch was inflexible in cases where a different
sub-project would like to use the Dear ImGui Vulkan backend but not use the
custom GITS Vulkan headers. Now this can be controlled by setting the
IMGUI_USE_GITS_VULKAN_HEADERS define.

This also removes the shared backend library since we want to compile
the files differently on case-by-case basis.

Also adds the ignore-whitespace argument to patch applying. Helps with
certain setups where the line endings differ (especially when working in
WSL). I believe this to be fine and safe for our use case.